hyprland: fixed click notification center making every other click be for notifications
https://github.com/Alexays/Waybar/issues/1968
https://github.com/Alexays/Waybar/pull/2414

It seems like running a process which starts a window on click causes
a bug where ever future click on the bar will open this program. The
solution seems to be to run the program when the click is released.

The comment in the issue says to use `on-click-released` field but
this does not exist. Instead as we can see in the PR the field is
named `on-click-release`.